cloudfunction/requirements.txt
google-cloud-videointelligence==2.0.0
google-cloud-storage==1.37.0
moviepy==1.0.3
GrabzIt==3.4.5.1
algoliasearch==2.4.0 
